requirements.txt
autopep8==1.5.6
Django==3.2
django-cors-headers==3.6.0
django-crispy-forms==1.14.0
djangorestframework==3.12.2
gunicorn==20.0.4
openpyxl==3.0.9
Pillow==8.4.0
pdfkit==1.0.0
psycopg2==2.8.6
python-dateutil==2.8.2
python-dotenv==0.15.0
